test(web-app-demo-backend): Add a test involving multiple chats
I know there still is a bug in there, because not wrapping
the `DashMap`s in `Arc`s is wrong. To lure it out, I
will have to create receivers without creating the chats,
than we should see defective histories, as they will be created
in the `ChatServers` cloned for the sending tasks and not
in the original test `ChatServer`.